# python-flask-server-generated/swagger_server/models/base_user_info.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class BaseUserInfo(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, password: str=None, token: str=None):  # noqa: E501
        """BaseUserInfo - a model defined in Swagger

        :param id: The id of this BaseUserInfo.  # noqa: E501
        :type id: str
        :param password: The password of this BaseUserInfo.  # noqa: E501
        :type password: str
        :param token: The token of this BaseUserInfo.  # noqa: E501
        :type token: str
        """
        self.swagger_types = {
            'id': str,
            'password': str,
            'token': str
        }

        self.attribute_map = {
            'id': 'id',
            'password': 'password',
            'token': 'token'
        }
        self._id = id
        self._password = password
        self._token = token

    @classmethod
    def from_dict(cls, dikt) -> 'BaseUserInfo':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BaseUserInfo of this BaseUserInfo.  # noqa: E501
        :rtype: BaseUserInfo
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this BaseUserInfo.


        :return: The id of this BaseUserInfo.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this BaseUserInfo.


        :param id: The id of this BaseUserInfo.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def password(self) -> str:
        """Gets the password of this BaseUserInfo.


        :return: The password of this BaseUserInfo.
        :rtype: str
        """
        return self._password

    @password.setter
    def password(self, password: str):
        """Sets the password of this BaseUserInfo.


        :param password: The password of this BaseUserInfo.
        :type password: str
        """
        if password is None:
            raise ValueError("Invalid value for `password`, must not be `None`")  # noqa: E501

        self._password = password

    @property
    def token(self) -> str:
        """Gets the token of this BaseUserInfo.


        :return: The token of this BaseUserInfo.
        :rtype: str
        """
        return self._token

    @token.setter
    def token(self, token: str):
        """Sets the token of this BaseUserInfo.


        :param token: The token of this BaseUserInfo.
        :type token: str
        """

        self._token = token