# python-flask-server-generated/swagger_server/models/error_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class ErrorResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, code: str=None, text: str=None):  # noqa: E501
        """ErrorResponse - a model defined in Swagger

        :param code: The code of this ErrorResponse.  # noqa: E501
        :type code: str
        :param text: The text of this ErrorResponse.  # noqa: E501
        :type text: str
        """
        self.swagger_types = {
            'code': str,
            'text': str
        }

        self.attribute_map = {
            'code': 'code',
            'text': 'text'
        }
        self._code = code
        self._text = text

    @classmethod
    def from_dict(cls, dikt) -> 'ErrorResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ErrorResponse of this ErrorResponse.  # noqa: E501
        :rtype: ErrorResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def code(self) -> str:
        """Gets the code of this ErrorResponse.


        :return: The code of this ErrorResponse.
        :rtype: str
        """
        return self._code

    @code.setter
    def code(self, code: str):
        """Sets the code of this ErrorResponse.


        :param code: The code of this ErrorResponse.
        :type code: str
        """
        if code is None:
            raise ValueError("Invalid value for `code`, must not be `None`")  # noqa: E501

        self._code = code

    @property
    def text(self) -> str:
        """Gets the text of this ErrorResponse.


        :return: The text of this ErrorResponse.
        :rtype: str
        """
        return self._text

    @text.setter
    def text(self, text: str):
        """Sets the text of this ErrorResponse.


        :param text: The text of this ErrorResponse.
        :type text: str
        """
        if text is None:
            raise ValueError("Invalid value for `text`, must not be `None`")  # noqa: E501

        self._text = text