# python-flask-server-generated/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Sports betting
  version: 0.1.9
servers:
- url: http://127.0.0.1:5000
  description: "Optional server description, e.g. Main (production) server"
- url: http://127.0.0.1:5000
  description: "Optional server description, e.g. Internal staging server for testing"
paths:
  /user_register:
    post:
      description: Регистрирует пользователя с заданными данными.
      operationId: user_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
      responses:
        "200":
          description: Удачная попытка регистрации
        "400":
          description: "Ошибка при валидации параметров (например, попытка повторн\
            ой регистрации)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /user_login:
    post:
      description: Залогинивает пользователя с заданными данными.
      operationId: user_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseUserInfo'
      responses:
        "200":
          description: Удачная попытка логина
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /user_register_approve:
    post:
      description: Подтверждает регистрацию пользователя (страница с вводом слова
        регистрации).
      operationId: user_register_approve_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterApprove'
      responses:
        "200":
          description: Удачное подтверждение
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Неверное кодовое слово
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /user:
    get:
      description: Выдача информации о пользователе
      operationId: user_get
      parameters:
      - name: id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    put:
      description: Изменение информации о пользователе.
      operationId: user_put
      parameters:
      - name: id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserMeta'
      responses:
        "200":
          description: Удачное изменение
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /competitions:
    get:
      description: Получает все соревнования пользователя.
      operationId: competitions_get
      parameters:
      - name: id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /competitions_info:
    get:
      description: Отдает информаицю о соревновании для этого пользовтеля.
      operationId: competitions_info_get
      parameters:
      - name: competition_id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь или соревнование не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      description: Добавляет пользователя в участие в соревнование.
      operationId: competitions_info_post
      parameters:
      - name: competition_id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь или соревнование не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /match_info:
    get:
      description: Отдает информаицю о матче для этого пользовтеля.
      operationId: match_info_get
      parameters:
      - name: match_id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь или соревнование не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /create/competition:
    post:
      description: Создает новое соревнование.
      operationId: create_competition_post
      parameters:
      - name: id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_competition_body'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь или соревнование не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /create/match:
    post:
      description: Создает новый матч.
      operationId: create_match_post
      parameters:
      - name: id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: competion_id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_match_body'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь или соревнование не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /create/bet:
    post:
      description: Делает ставку.
      operationId: create_bet_post
      parameters:
      - name: id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: match_id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_bet_body'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        "400":
          description: Ошибка при валидации параметров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь или соревнование не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
components:
  schemas:
    UserMeta:
      required:
      - password
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        password:
          type: string
      example:
        password: password
        surname: surname
        name: name
    UserInfo:
      required:
      - id
      - user_meta
      type: object
      properties:
        user_meta:
          $ref: '#/components/schemas/UserMeta'
        id:
          type: string
      example:
        user_meta:
          password: password
          surname: surname
          name: name
        id: id
    BaseUserInfo:
      required:
      - id
      - password
      type: object
      properties:
        id:
          type: string
        password:
          type: string
        token:
          type: string
    RegisterApprove:
      required:
      - id
      - secret_code
      type: object
      properties:
        id:
          type: string
        secret_code:
          type: string
    ErrorResponse:
      required:
      - code
      - text
      type: object
      properties:
        code:
          type: string
        text:
          type: string
    Match:
      required:
      - first_team_name
      - id
      - is_active
      - name
      - second_team_name
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        first_team_name:
          type: string
        second_team_name:
          type: string
        first_team_result:
          type: integer
        second_team_result:
          type: integer
        is_active:
          type: boolean
        bets_result:
          type: number
        user_bets:
          type: array
          items:
            $ref: '#/components/schemas/Bets'
      example:
        user_bets:
        - bet: 2
          name: name
          bet_profit: 7.061401241503109
        - bet: 2
          name: name
          bet_profit: 7.061401241503109
        second_team_result: 5
        bets_result: 5.637376656633329
        name: name
        id: 6
        is_active: true
        first_team_result: 1
        first_team_name: first_team_name
        second_team_name: second_team_name
    Competition:
      required:
      - id
      - is_active
      - matches
      - name
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
        is_active:
          type: boolean
        matches:
          type: array
          items:
            $ref: '#/components/schemas/Match'
        leader_board:
          type: array
          items:
            $ref: '#/components/schemas/BetsResult'
      example:
        is_active: true
        name: name
        id: 0
        matches:
        - user_bets:
          - bet: 2
            name: name
            bet_profit: 7.061401241503109
          - bet: 2
            name: name
            bet_profit: 7.061401241503109
          second_team_result: 5
          bets_result: 5.637376656633329
          name: name
          id: 6
          first_team_result: 1
          first_team_name: first_team_name
          second_team_name: second_team_name
        - user_bets:
          - bet: 2
            name: name
            bet_profit: 7.061401241503109
          - bet: 2
            name: name
            bet_profit: 7.061401241503109
          second_team_result: 5
          bets_result: 5.637376656633329
          name: name
          id: 6
          first_team_result: 1
          first_team_name: first_team_name
          second_team_name: second_team_name
        leader_board:
        - result: 9.301444243932576
          user_id: user_id
          bets:
          - null
          - null
        - result: 9.301444243932576
          user_id: user_id
          bets:
          - null
          - null
    Bets:
      required:
      - bet
      - name
      type: object
      properties:
        name:
          type: string
        bet:
          type: integer
        bet_profit:
          type: number
      example:
        bet: 2
        name: name
        bet_profit: 7.061401241503109
    BetsResult:
      required:
      - result
      - user_id
      type: object
      properties:
        user_id:
          type: string
        result:
          type: number
        bets:
          type: array
          items:
            $ref: '#/components/schemas/Bets'
      example:
        result: 9.301444243932576
        user_id: user_id
        bets:
        - null
        - null
    inline_response_200:
      required:
      - competitions
      type: object
      properties:
        competitions:
          type: array
          items:
            $ref: '#/components/schemas/Competition'
      example:
        competitions:
        - is_active: true
          name: name
          id: 0
          matches:
          - user_bets:
            - bet: 2
              name: name
              bet_profit: 7.061401241503109
            - bet: 2
              name: name
              bet_profit: 7.061401241503109
            second_team_result: 5
            bets_result: 5.637376656633329
            name: name
            id: 6
            first_team_result: 1
            first_team_name: first_team_name
            second_team_name: second_team_name
          - user_bets:
            - bet: 2
              name: name
              bet_profit: 7.061401241503109
            - bet: 2
              name: name
              bet_profit: 7.061401241503109
            second_team_result: 5
            bets_result: 5.637376656633329
            name: name
            id: 6
            first_team_result: 1
            first_team_name: first_team_name
            second_team_name: second_team_name
          leader_board:
          - result: 9.301444243932576
            user_id: user_id
            bets:
            - null
            - null
          - result: 9.301444243932576
            user_id: user_id
            bets:
            - null
            - null
        - is_active: true
          name: name
          id: 0
          matches:
          - user_bets:
            - bet: 2
              name: name
              bet_profit: 7.061401241503109
            - bet: 2
              name: name
              bet_profit: 7.061401241503109
            second_team_result: 5
            bets_result: 5.637376656633329
            name: name
            id: 6
            first_team_result: 1
            first_team_name: first_team_name
            second_team_name: second_team_name
          - user_bets:
            - bet: 2
              name: name
              bet_profit: 7.061401241503109
            - bet: 2
              name: name
              bet_profit: 7.061401241503109
            second_team_result: 5
            bets_result: 5.637376656633329
            name: name
            id: 6
            first_team_result: 1
            first_team_name: first_team_name
            second_team_name: second_team_name
          leader_board:
          - result: 9.301444243932576
            user_id: user_id
            bets:
            - null
            - null
          - result: 9.301444243932576
            user_id: user_id
            bets:
            - null
            - null
    create_competition_body:
      required:
      - parsing_ref
      type: object
      properties:
        parsing_ref:
          type: string
    create_match_body:
      required:
      - parsing_ref
      type: object
      properties:
        parsing_ref:
          type: string
    create_bet_body:
      required:
      - bets
      type: object
      properties:
        bets:
          type: array
          items:
            $ref: '#/components/schemas/Bets'
    inline_response_200_1:
      required:
      - token
      type: object
      properties:
        token:
          type: string
      example:
        token: token